Created lettings app and temporary tests.
