+++ lettings/fixtures/lettings_data.json
@@ -1,0 +1,1 @@
+[{"model": "lettings.address", "pk": 1, "fields": {"number": 7217, "street": "Bedford Street", "city": "Brunswick", "state": "GA", "zip_code": 31525, "country_iso_code": "USA"}}, {"model": "lettings.address", "pk": 2, "fields": {"number": 4, "street": "Military Street", "city": "Willoughby", "state": "OH", "zip_code": 44094, "country_iso_code": "USA"}}, {"model": "lettings.address", "pk": 3, "fields": {"number": 340, "street": "Wintergreen Avenue", "city": "Newport News", "state": "VA", "zip_code": 23601, "country_iso_code": "USA"}}, {"model": "lettings.address", "pk": 4, "fields": {"number": 9230, "street": "E. Joy Ridge Street", "city": "Marquette", "state": "MI", "zip_code": 49855, "country_iso_code": "USA"}}, {"model": "lettings.address", "pk": 5, "fields": {"number": 9606, "street": "Harvard Street", "city": "Aliquippa", "state": "PA", "zip_code": 15001, "country_iso_code": "USA"}}, {"model": "lettings.address", "pk": 6, "fields": {"number": 588, "street": "Argyle Avenue", "city": "East Meadow", "state": "NY", "zip_code": 11554, "country_iso_code": "USA"}}, {"model": "lettings.letting", "pk": 1, "fields": {"title": "Joshua Tree Green Haus /w Hot Tub", "address": 1}}, {"model": "lettings.letting", "pk": 2, "fields": {"title": "Oceanview Retreat", "address": 2}}, {"model": "lettings.letting", "pk": 3, "fields": {"title": "'Silo Studio' Cottage", "address": 3}}, {"model": "lettings.letting", "pk": 4, "fields": {"title": "Pirates of the Caribbean Getaway", "address": 4}}, {"model": "lettings.letting", "pk": 5, "fields": {"title": "The Mushroom Dome Retreat & LAND of Paradise Suite", "address": 5}}, {"model": "lettings.letting", "pk": 6, "fields": {"title": "Underground Hygge", "address": 6}}]
