+++ oc_lettings_site/admin.py
@@ -1,8 +1,1 @@
 from django.contrib import admin
-
-from .models import Letting
-from .models import Address
-
-
-admin.site.register(Letting)
-admin.site.register(Address)
