+++ oc_lettings_site/urls.py
@@ -6,7 +6,6 @@ from . import views
 urlpatterns = [
     path('', views.index, name='index'),
     path('', include('profiles.urls')),
-    path('lettings/', views.lettings_index, name='lettings_index'),
-    path('lettings/<int:letting_id>/', views.letting, name='letting'),
+    path('', include('lettings.urls')),
     path('admin/', admin.site.urls),
 ]
